volleyball-coach/requirements.txt
langchain==0.1.4
langchain-openai==0.0.5
opencv-python==4.8.1.78
flask==2.3.3
google-generativeai==0.3.2
pillow==10.1.0
python-dotenv==1.0.0
numpy==1.26.0
requests==2.31.0 